pyproject.toml
[project]
name = "jira-ai-agent"
version = "0.1.0"
description = "JIRA AI Agent - Resolves Product and Technical Queries by analyzing historical JIRA tickets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.2.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "streamlit>=1.31.0",
    "python-dotenv>=1.0.0",
    "jira>=3.8.0",
    "requests>=2.31.0",
    "pandas>=2.2.0",
    "plotly>=5.18.0",
    "python-dateutil>=2.8.2",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
